requirements.txt
streamlit
matplotlib
pandas
networkx
python-docx
psutil